// src/common/enums/user.enum.ts
export enum EUserRole {
  Admin = 'admin',
  User = 'user',
  Staff = 'staff',
}
export enum EUserStatus {
  ACTIVE = 'active',
  INACTIVE = 'inactive',
}

export enum ESeatStatus {
  AVAILABLE = 'available',
  RESERVED = 'reserved',
  SELECTED = 'selected',
  WAITING = 'waiting',
  OPEN = 'active',
  CLOSE = 'close',
}

export enum ETicketStatus {
  ACTIVE = 'active',
  INACTIVE = 'inactive',
  COMPLETE = 'complete',
}

export enum EMdTicket {
  MDTICKET = 'mdticket',
}

export enum ENotificationType {
  SINGLEUSER = 'singleuser',
  ALLUSER = 'alluser',
}

export enum ENotificationStatus {
  ACTIVE = 'active',
  INACTIVE = 'inactive',
}

export enum EDiscountStatus {
  ACTIVE = 'active',
  INACTIVE = 'inactive',
}